Share the same PhantomSQL object among all requests
With this commit the same SimpleSQL object is used by all requests. This
object creates one PhantomSQL object, which in turns corresponds to a
single SQLAlchemy session. We remove the reset_db decorator so that the
PhantomSQL object stays the same.

According to SQLAlchemy documentation, this is not supposed to work as
session objects should be per-thread, not global.